[m-organize-table-columns] Bugfix/Probleme de import dans le MOrganizeTableColumns (#258)
* fix import+ ajout regle tslint

* fix regex
